requirements.txt
asgiref==3.8.1
Django==5.1.2
django-jazzmin==3.0.1
django-modeltranslation==0.19.10
djangorestframework==3.15.2
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.2
